lib/python/requirements.txt
colorama==0.3.6
sh==1.11
requests==2.11.1
invoke==0.13.0
numpy==1.11.1
PyYAML==3.12
flake8==3.0.4
autopep8==1.2.4
